+++ apps/server/src/app/apollo/error-formatter.ts
