+++ apps/server/src/app/apollo/mappers/event.mapper.ts
