+++ apps/server/src/app/apollo/mappers/program.mapper.ts
