+++ apps/server/src/app/apollo/mappers/registration.mapper.ts
